refactor(smart-contract): separe logic into hpp and cpp
