+++ contracts/scripts/compile.sh
@@ -1,0 +1,3 @@
+echo 'Compiling rateproducer Smart Contract'
+
+eosio-cpp -abigen -I ../rateproducer/include -R ../rateproducer/ricardian -contract rateproducer -o rateproducer.wasm ../rateproducer/src/rateproducer.cpp
